+++ contracts/chain-id-mappings/LayerZeroChainIdMapping.sol
@@ -28,42 +28,29 @@ contract LayerZeroChainIdMapping is Ownable2Step {
     mapping(uint16 => uint64) public upMapping;
 
     constructor(uint64[] memory _lineRegistryChainIds, uint16[] memory _lzChainIds) {
-        require(
-            _lineRegistryChainIds.length == _lzChainIds.length,
-            "Lengths do not match."
-        );
+        require(_lineRegistryChainIds.length == _lzChainIds.length, "Lengths do not match.");
 
-        for (uint i = 0; i < _lineRegistryChainIds.length; i++) {
+        for (uint256 i = 0; i < _lineRegistryChainIds.length; i++) {
             downMapping[_lineRegistryChainIds[i]] = _lzChainIds[i];
             upMapping[_lzChainIds[i]] = _lineRegistryChainIds[i];
         }
     }
 
-    function addChainIdMap(
-        uint64 _lineRegistryChainId,
-        uint16 _lzChainId
-    ) external onlyOwner {
-        require(
-            downMapping[_lineRegistryChainId] == 0,
-            "LineRegistryChainId already exists."
-        );
+    function addChainIdMap(uint64 _lineRegistryChainId, uint16 _lzChainId) external onlyOwner {
+        require(downMapping[_lineRegistryChainId] == 0, "LineRegistryChainId already exists.");
         require(upMapping[_lzChainId] == 0, "lzChainId already exists.");
         downMapping[_lineRegistryChainId] = _lzChainId;
         upMapping[_lzChainId] = _lineRegistryChainId;
     }
 
-    function down(
-        uint64 lineRegistryChainId
-    ) external view returns (uint16 lzChainId) {
+    function down(uint64 lineRegistryChainId) external view returns (uint16 lzChainId) {
         lzChainId = downMapping[lineRegistryChainId];
         if (lzChainId == 0) {
             revert LineRegistryChainIdNotFound(lineRegistryChainId);
         }
     }
 
-    function up(
-        uint16 lzChainId
-    ) external view returns (uint64 lineRegistryChainId) {
+    function up(uint16 lzChainId) external view returns (uint64 lineRegistryChainId) {
         lineRegistryChainId = upMapping[lzChainId];
         if (lineRegistryChainId == 0) {
             revert LineRegistryChainIdNotFound(lineRegistryChainId);
